requirements.txt
streamlit==1.10.0  # ou une version plus récente
requests==2.28.1
beautifulsoup4==4.11.1
Pillow==9.1.1